feat: add settings config and .env

+++ src/meridian_backend/models/product.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-
+from decimal import Decimal
 from uuid import UUID
 
 
@@ -7,4 +7,4 @@ from uuid import UUID
 class Product:
     id: UUID
     name: str
-    price: float
+    price: Decimal
